Add simple request handler with config
We are going to use the `sawyer` gem for communicating with the
google host, but let's a simple wrapper on top of it so, we can
simplify each of our request.

This commit also adds an API configuration option, which can be
use to add application specific API key.

```ruby
Googlemap.configuration.api_key = "YOUR-SECRET-API_KEY"
```