Interface to find an address geocode
This commit adds the `Googlemap::Geocode` interface, that expose
the `find` method which we can use to retrieve the details for any
valid address.

For valid address it should returns the details in `results`, but
if the address is invalid or something else then it will return an
empty `results` with `ZERO_RESULTS` in the status.

```ruby
Googlemap::Geocode.find("valid_address")
```